Rename files to match class to filename
