+++ src/app/clusters/valve-configuration-and-control-server/BUILD.gn
@@ -17,8 +17,8 @@ import("//build_overrides/chip.gni")
 source_set("valve-configuration-and-control-server") {
   sources = [
     "TimeSyncTracker.h",
-    "valve-configuration-and-control-cluster.cpp",
-    "valve-configuration-and-control-cluster.h",
+    "ValveConfigurationAndControlCluster.cpp",
+    "ValveConfigurationAndControlCluster.h",
     "valve-configuration-and-control-delegate.h",
   ]
